fund-portfolio-intelligence/requirements.txt
# ============================================
# Fund Portfolio Intelligence System
# Python Dependencies
# ============================================

# Core ML/AI Frameworks
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
langgraph>=0.0.20
langchain-neo4j>=0.0.5
dspy-ai>=2.4.0

# LLM Providers
groq>=0.4.0

# Azure Services
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0

# Database Drivers
neo4j>=5.14.0
pyprojroot>=0.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration Management
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# API & Web
yfinance>=0.2.30
requests>=2.31.0

# Utilities
tabulate>=0.9.0  # For markdown table support
pathlib>=1.0.1
typing-extensions>=4.8.0

# Development Tools (Optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
ipython>=8.12.0

# Logging & Monitoring
colorlog>=6.7.0
